Very rough MirrorBag; Arrangement changed name
Arrangement is now MirrorSet, which better reflects what it does
(mirrors the relationship between keys to positions, and positions to
keys). MirrorBag is a counterpart that allows potentially-multiple
duplicate items instead of the set-like requirement for unique items; it
is at an extremely early state and is likely to not work completely yet.